# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HackR API",
        "description": "Here is my Swagger documentation for my HackR API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/logs/recent": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Get the last logs",
                "operationId": "3898208f01f5053eba7778f231537d2a",
                "parameters": [
                    {
                        "name": "nbLogs",
                        "in": "query",
                        "description": "Number of logs to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of the last logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "date_action": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-01-14T15:04:05"
                                            },
                                            "user": {
                                                "type": "string",
                                                "example": "Haitam Elqassimi"
                                            },
                                            "fonctionnalite": {
                                                "type": "string",
                                                "example": "Email spam"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous n'avez pas le droit pour faire cela, seul l'admin peut le faire."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while retrieving logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de la récupération des logs"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logs/user": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Get logs by user email",
                "operationId": "3f5d2b23875b8500a36ccfd6710199c8",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of the user to retrieve logs for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "user@example.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of logs for the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "date_action": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-01-14T15:04:05"
                                            },
                                            "user": {
                                                "type": "string",
                                                "example": "Haitam Elqassimi",
                                                "nullable": true
                                            },
                                            "fonctionnalite": {
                                                "type": "string",
                                                "example": "Email spam",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email parameter is missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Paramètre manquant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous n'avez pas le droit pour faire cela, seul l'admin peut le faire."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while retrieving logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de la récupération des logs"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logs/functionality/{nom_fonctionnalite}": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Get logs by functionality",
                "operationId": "d72119f04032e42d2bbef132a98dabe1",
                "parameters": [
                    {
                        "name": "nom_fonctionnalite",
                        "in": "path",
                        "description": "Name of the functionality to retrieve logs for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Email spam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of logs for the functionality",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "date_action": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-01-14T15:04:05"
                                            },
                                            "user": {
                                                "type": "string",
                                                "example": "Haitam Elqassimi"
                                            },
                                            "fonctionnalite": {
                                                "type": "string",
                                                "example": "Email spam"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous n'avez pas le droit pour faire cela, seul l'admin peut le faire."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Functionality not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Fonctionnalité non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while retrieving logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur lors de la récupération des logs"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}